chore(prot): change credits represented as bigInt's to uint64

+++ protocol/lite_data_entry_test.go
@@ -63,7 +63,7 @@ func TestLiteDataEntry(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
-	if cost != FeeWriteData.AsInt() {
+	if cost != FeeWriteData.AsUInt64() {
 		t.Fatalf("expected a cost of 10 credits, but computed %d", cost)
 	}
 
@@ -80,7 +80,7 @@ func TestLiteDataEntry(t *testing.T) {
 	}
 
 	//the size is now 987 bytes so it should cost 40 credits
-	if cost != 4*FeeWriteData.AsInt() {
+	if cost != 4*FeeWriteData.AsUInt64() {
 		t.Fatalf("expected a cost of 40 credits, but computed %d", cost)
 	}
